PB-934: Avoid python cyclic import
New pylint doesn't allow anymore cyclic import, solve it using a method found
in https://stackoverflow.com/questions/48718768/how-to-avoid-circular-imports-in-a-flask-app-with-flask-sqlalchemy-models
